+++ src/definition/populate.go
@@ -16,11 +16,12 @@ const (
 type GuessFunc func(groupLabel string) (string, error)
 
 type IntegrationPopulateConfig struct {
-	Integration   *integration.Integration
-	ClusterName   string
-	K8sVersion    fmt.Stringer
-	MsTypeGuesser GuessFunc
-	Groups        RawGroups
-	Specs         SpecGroups
-	Filterer      discovery.NamespaceFilterer
+	Integration    *integration.Integration
+	ClusterName    string
+	CloudClusterID string
+	K8sVersion     fmt.Stringer
+	MsTypeGuesser  GuessFunc
+	Groups         RawGroups
+	Specs          SpecGroups
+	Filterer       discovery.NamespaceFilterer
 }