+++ src/ksm/scraper.go
@@ -19,9 +19,11 @@ import (
 	"github.com/newrelic/nri-kubernetes/v3/src/scrape"
 )
 
-const defaultLabelSelector = "app.kubernetes.io/name=kube-state-metrics"
-const defaultScheme = "http"
-const ksmMetricsPath = "metrics"
+const (
+	defaultLabelSelector = "app.kubernetes.io/name=kube-state-metrics"
+	defaultScheme        = "http"
+	ksmMetricsPath       = "metrics"
+)
 
 // Providers is a struct holding pointers to all the clients Scraper needs to get data from.
 // TODO: Extract this out of the KSM package.
@@ -35,6 +37,7 @@ type Scraper struct {
 	Providers
 	logger              *log.Logger
 	config              *config.Config
+	cloudClusterID      string
 	k8sVersion          *version.Info
 	endpointsDiscoverer discovery.EndpointsDiscoverer
 	servicesLister      listersv1.ServiceLister
@@ -42,7 +45,7 @@ type Scraper struct {
 	Filterer            discovery.NamespaceFilterer
 }
 
-// ScraperOpt are options that can be used to configure the Scraper
+// ScraperOpt are options that can be used to configure the Scraper.
 type ScraperOpt func(s *Scraper) error
 
 // WithLogger returns an OptionFunc to change the logger from the default noop logger.
@@ -57,6 +60,14 @@ func WithLogger(logger *log.Logger) ScraperOpt {
 func WithFilterer(filterer discovery.NamespaceFilterer) ScraperOpt {
 	return func(s *Scraper) error {
 		s.Filterer = filterer
+		return nil
+	}
+}
+
+// WithCloudClusterID returns an OptionFunc to set the cloud-detected cluster id.
+func WithCloudClusterID(id string) ScraperOpt {
+	return func(s *Scraper) error {
+		s.cloudClusterID = id
 		return nil
 	}
 }
@@ -134,7 +145,7 @@ func (s *Scraper) Run(i *integration.Integration) error {
 		job := scrape.NewScrapeJob("kube-state-metrics", grouper, metric.KSMSpecs, scrape.JobWithFilterer(s.Filterer))
 
 		s.logger.Debugf("Running KSM job")
-		r := job.Populate(i, s.config.ClusterName, s.logger, s.k8sVersion)
+		r := job.Populate(i, s.config.ClusterName, s.cloudClusterID, s.logger, s.k8sVersion)
 		if r.Errors != nil {
 			if r.Populated {
 				s.logger.Tracef("Error populating KSM metrics: %v", r.Error())