+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### enhancement
+- Automatically detect cluster resource ids (EKS ARN, AKS ARM id, or GKE link). Configurable with `common.config.disableCloudClusterIdDetection: true/false`
+  @dbudziwojski [#1520](https://github.com/newrelic/nri-kubernetes/pull/1520)
+
 ### security
 - Reverts #1518 that prevented scanners from flagging CVE-2013-3900 on this image but did not add support for WinVerifyTrust. @dbudziwojski [#1524](https://github.com/newrelic/nri-kubernetes/pull/1524)